PPU_FIIT_AR-59

See merge request urfu2022/fiit/tutoraggregator!80

+++ SPA/V1/DataModels/V1LessonDto.cs
@@ -1,32 +1,31 @@
 ﻿using Newtonsoft.Json;
+using SPA.Domain;
 
 namespace SPA.V1.DataModels;
 
-using Domain;
-
 public sealed class V1LessonDto
 {
-    [JsonProperty("id")]
+    [JsonProperty("id")] 
     public string Id { get; init; }
-    
-    [JsonProperty("price")]
-    public double Price { get; init; }
-    
-    [JsonProperty("status")]
+
+    [JsonProperty("price")] 
+    public int Price { get; init; }
+
+    [JsonProperty("status")] 
     public LessonStatus Status { get; init; }
-    
-    [JsonProperty("type")]
+
+    [JsonProperty("type")] 
     public LessonType Type { get; init; }
-    
-    [JsonProperty("start")]
+
+    [JsonProperty("start")] 
     public DateTimeOffset Start { get; init; }
-    
-    [JsonProperty("end")]
+
+    [JsonProperty("end")] 
     public DateTimeOffset End { get; init; }
 
-    [JsonProperty("tutor")]
+    [JsonProperty("tutor")] 
     public V1TutorInfoDto Tutor { get; init; }
-    
-    [JsonProperty("student")]
+
+    [JsonProperty("student")] 
     public V1StudentInfoDto Student { get; init; }
 }
